+++ src/main/java/org/moltimate/moltimatebackend/controller/AlignmentController.java
@@ -5,8 +5,8 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import lombok.extern.slf4j.Slf4j;
-import org.moltimate.moltimatebackend.dto.ActiveSiteAlignmentRequest;
-import org.moltimate.moltimatebackend.dto.ActiveSiteAlignmentResponse;
+import org.moltimate.moltimatebackend.dto.ActiveSiteAlignment.ActiveSiteAlignmentRequest;
+import org.moltimate.moltimatebackend.dto.ActiveSiteAlignment.ActiveSiteAlignmentResponse;
 import org.moltimate.moltimatebackend.service.AlignmentService;
 import org.moltimate.moltimatebackend.exception.InvalidPdbIdException;
 import org.springframework.beans.factory.annotation.Autowired;
