+++ src/main/java/org/moltimate/moltimatebackend/dto/ActiveSiteAlignment/ActiveSiteAlignmentRequest.java
@@ -1,8 +1,10 @@
-package org.moltimate.moltimatebackend.dto;
+package org.moltimate.moltimatebackend.dto.ActiveSiteAlignment;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.moltimate.moltimatebackend.dto.MotifFile;
+import org.moltimate.moltimatebackend.dto.PdbQueryResponse;
 import org.moltimate.moltimatebackend.model.Motif;
 import org.moltimate.moltimatebackend.util.FileUtils;
 import org.moltimate.moltimatebackend.util.ProteinUtils;
