+++ src/main/java/org/moltimate/moltimatebackend/dto/MotifTesting/SuccessfulAlignment.java
