+++ src/main/java/org/moltimate/moltimatebackend/service/MotifTestService.java
@@ -3,7 +3,7 @@ package org.moltimate.moltimatebackend.service;
 import lombok.extern.slf4j.Slf4j;
 import org.biojava.nbio.structure.Structure;
 import org.moltimate.moltimatebackend.dto.MotifFile;
-import org.moltimate.moltimatebackend.dto.MotifTestRequest;
+import org.moltimate.moltimatebackend.dto.MotifTesting.MotifTestRequest;
 import org.moltimate.moltimatebackend.dto.MotifTesting.FailedAlignment;
 import org.moltimate.moltimatebackend.dto.MotifTesting.MotifTestResponse;
 import org.moltimate.moltimatebackend.dto.MotifTesting.SuccessfulAlignment;
